refactor(ignitionserver): move lookupMappedImage to shared util package
Move lookupMappedImage function from ignitionserver deployment to support/util/imagemapping.go as LookupMappedImage.
Sync changes from PR #6517:
- https://github.com/openshift/hypershift/pull/6517/commits/80b31db9b30e63201a13ac37bb6cc4ab8c5cfd55
- https://github.com/openshift/hypershift/pull/6517#discussion_r2266232699"
